fix rubocop errors. Fix pagination error

+++ app/controllers/charges_controller.rb
@@ -1,24 +1,14 @@
 # Charges controller
 class ChargesController < ApplicationController
   include SessionHelper
+  before_action :set_current_order
+  before_action :set_amount
+  require 'date'
+
   def create
-    @order = current_order
-    @amount = ((@order.subtotal + (@order.subtotal * (@order.pst + @order.gst +
-                @order.hst))) * 100).to_i
     @customer = new_stripe_customer
     @charge = new_strip_charge
-
-    require 'date'
-    # save stripe customer id with stripe customer!
-    if @charge.paid
-      @order = current_order
-      @order.order_date = DateTime.now
-      @order.customer_id = current_customer.id
-      @order.order_status_id = 2
-      @order.save
-      session.delete(:order_id)
-      # current_order = Order.new
-    end
+    session.delete(:order_id) if @charge.paid
 
   rescue Stripe::CardError => e
     flash[:error] = e.message
@@ -35,5 +25,22 @@ class ChargesController < ApplicationController
                           amount: (@amount * 100).to_i,
                           description: 'Rails Stripe Customer',
                           currency: 'cad')
+  end
+
+  def update_order
+    @order = current_order
+    @order.order_date = DateTime.current
+    @order.customer_id = current_customer.id
+    @order.order_status_id = 2
+    @order.save
+  end
+
+  def set_current_order
+    @order = current_order
+  end
+
+  def set_amount
+    @amount = ((@order.subtotal + (@order.subtotal * (@order.pst + @order.gst +
+                @order.hst))) * 100).to_i
   end
 end
